+++ packages/natives/native/index.js
@@ -24,13 +24,14 @@ export const PtySession = nativeBindings.PtySession;
 export const Shell = nativeBindings.Shell;
 
 // functions
-export const __piNativesV0_3_1 = nativeBindings.__piNativesV0_3_1;
+export const __piNativesV0_3_2 = nativeBindings.__piNativesV0_3_2;
 export const applyBashFixups = nativeBindings.applyBashFixups;
 export const astEdit = nativeBindings.astEdit;
 export const astGrep = nativeBindings.astGrep;
 export const copyToClipboard = nativeBindings.copyToClipboard;
 export const countTokens = nativeBindings.countTokens;
 export const detectMacOSAppearance = nativeBindings.detectMacOSAppearance;
+export const diffLines = nativeBindings.diffLines;
 export const encodeSixel = nativeBindings.encodeSixel;
 export const executeShell = nativeBindings.executeShell;
 export const extractSegments = nativeBindings.extractSegments;
@@ -39,6 +40,9 @@ export const getSupportedLanguages = nativeBindings.getSupportedLanguages;
 export const getWorkProfile = nativeBindings.getWorkProfile;
 export const glob = nativeBindings.glob;
 export const grep = nativeBindings.grep;
+export const h01FindBestFuzzyMatch = nativeBindings.h01FindBestFuzzyMatch;
+export const h02ScoreSequenceFuzzy = nativeBindings.h02ScoreSequenceFuzzy;
+export const h06FormatHashLines = nativeBindings.h06FormatHashLines;
 export const hasMatch = nativeBindings.hasMatch;
 export const highlightCode = nativeBindings.highlightCode;
 export const htmlToMarkdown = nativeBindings.htmlToMarkdown;